+++ src/main/java/com/example/javabackend/dto/StockDTO.java
@@ -1,17 +1,19 @@
 package com.example.javabackend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class StockDTO {
     private String symbol;
     private String name;
-    private Double price;
-    private Double change;
-    private Double changePercent;
-    private Double marketCap;
-    private Double high;
-    private Double low;
-    private Double open;
-    private Double previousClose;
+    private BigDecimal price;
+    private BigDecimal change;
+    private BigDecimal changePercent;
+    private BigDecimal marketCap;
+    private BigDecimal high;
+    private BigDecimal low;
+    private BigDecimal open;
+    private BigDecimal previousClose;
 }
